Fixes file saving in `formatData`.
Previously, `formatData`  did not correctly handle cases when the
returned content was binary in nature but was not a `Buffer` instance.
This led to bugs such as failures when running:

`npx webpagetest waterfall <test_id> -o out.png -k <key>`

I'm not sure this patch is 100% correct, as the tests require Docker and
that isn't running on my current device.